refactor: userVideos.vue -> UserVideos.vue
