+++ src/front-end/src/views/My.vue
@@ -188,7 +188,7 @@
 
 <script>
 import { mapGetters, mapState } from 'vuex';
-import UserVideos from '@/components/userVideos';
+import UserVideos from '@/components/UserVideos';
 import { loadTossPayments } from '@tosspayments/payment-sdk';
 
 const clientKey = 'test_ck_ADpexMgkW36nWZAzQJE3GbR5ozO0';
